Moving time changing methods to EventTriggerer
