+++ NitroxServer/Communication/Packets/Processors/BedEnterProcessor.cs
@@ -1,5 +1,4 @@
-﻿using NitroxModel.Core;
-using NitroxModel.Packets;
+﻿using NitroxModel.Packets;
 using NitroxServer.Communication.Packets.Processors.Abstract;
 using NitroxServer.GameLogic;
 
@@ -7,16 +6,16 @@ namespace NitroxServer.Communication.Packets.Processors
 {
     public class BedEnterProcessor : AuthenticatedPacketProcessor<BedEnter>
     {
-        private readonly ScheduleKeeper scheduleKeeper;
+        private readonly EventTriggerer eventTriggerer;
 
-        public BedEnterProcessor(ScheduleKeeper scheduleKeeper)
+        public BedEnterProcessor(EventTriggerer eventTriggerer)
         {
-            this.scheduleKeeper = scheduleKeeper;
+            this.eventTriggerer = eventTriggerer;
         }
 
         public override void Process(BedEnter packet, Player player)
         {
-            scheduleKeeper.ChangeTime(ScheduleKeeper.TimeModification.SKIP);
+            eventTriggerer.ChangeTime(EventTriggerer.TimeModification.SKIP);
         }
     }
 }
